refactor(cli): extract config and savings rendering from commands/mod.rs
Continue splitting the command grab-bag: move cmd_config + its ConfigDisplay
types and format_* helpers into commands/config_display.rs, and the savings
report renderer (savings_scope_label, fmt_tokens, render_savings) into
commands/savings.rs. Tests moved with their subject. mod.rs re-exports
cmd_config; cmd_stats keeps using the savings helpers via a private use.

Pure code movement, no API change (all cartog tests pass). mod.rs is now
1591 lines, down from 3022 at the start of the split.
