+++ crates/cartog/src/commands/mod.rs
@@ -731,117 +731,8 @@ pub fn cmd_pull(
     remote::pull_index(db_path, config, cli_remote, force, no_sign_request, json)
 }
 
-/// Project label shown above the savings report. Strips the SQLite filename
-/// and walks up to the closest meaningful directory (typically the git root)
-/// so users see `cartog · my-project · 5 queries` rather than the full path.
-///
-/// Returns an empty string when nothing recognizable can be extracted (no
-/// parent, non-UTF-8 path component). The caller renders a header without a
-/// scope segment in that case — see [`render_savings`].
-fn savings_scope_label(db_path: &Path) -> String {
-    db_path
-        .parent()
-        .and_then(|p| {
-            // `.../<project>/.cartog/db.sqlite` → walk past `.cartog` to the
-            // project dir. `.../<project>/.cartog.db` → just take the parent.
-            if p.file_name().is_some_and(|n| n == ".cartog") {
-                p.parent()
-            } else {
-                Some(p)
-            }
-        })
-        .and_then(|p| p.file_name())
-        .and_then(|n| n.to_str())
-        .map(|n| n.to_string())
-        .unwrap_or_default()
-}
-
-/// Format a token count as `~1.4k`, `~28.5k`, etc. Matches the CCE-style
-/// compact display so the table stays narrow. Counts under 1,000 render as
-/// the raw integer.
-fn fmt_tokens(n: u64) -> String {
-    if n < 1_000 {
-        format!("{n}")
-    } else if n < 1_000_000 {
-        format!("~{:.1}k", n as f64 / 1_000.0)
-    } else {
-        format!("~{:.1}M", n as f64 / 1_000_000.0)
-    }
-}
-
-/// Render the savings report — header / bar / with-vs-without table / per-tool
-/// breakdown / footer. Pulled out of `cmd_stats` so the formatter is unit-
-/// testable independent of the DB.
-fn render_savings(scope: &str, r: &cartog_db::SavingsReport) -> String {
-    use cartog_db::{TOKENS_PER_QUERY_CARTOG, TOKENS_PER_QUERY_GREP};
-
-    let mut out = String::new();
-    if r.total_queries == 0 {
-        out.push_str(
-            "No queries logged yet. Run `cartog search`, `cartog refs`, … or \
-             point an MCP-aware editor at this index, then re-run.\n",
-        );
-        return out;
-    }
-
-    // Header: `cartog · my-project · N queries`
-    if scope.is_empty() {
-        out.push_str(&format!("cartog · {} queries\n\n", r.total_queries));
-    } else {
-        out.push_str(&format!(
-            "cartog · {scope} · {} queries\n\n",
-            r.total_queries
-        ));
-    }
-
-    // 10-cell bar with round-to-nearest. `percent_saved` is 0..=100; mapping
-    // `(p + 5) / 10` gives [0,4]→0 cells, [5,14]→1 cell, …, [95,100]→10
-    // cells, so the bar covers the full range. Previously `(p / 10).min(10)`
-    // capped at 9 cells (because `percent_saved` was clamped to 99), making
-    // 90% and 99% visually identical.
-    let filled = ((r.percent_saved as usize + 5) / 10).min(10);
-    let bar: String = "█".repeat(filled) + &"░".repeat(10 - filled);
-    out.push_str(&format!("{bar}  ~{}% tokens saved\n\n", r.percent_saved));
-
-    // With/without/saved table. Column widths chosen so the numbers line up
-    // for any plausible value: `Without cartog  ~999.9k tokens   (~1,700 / query)`.
-    let with = fmt_tokens(r.tokens_used_cartog);
-    let without = fmt_tokens(r.tokens_used_grep);
-    let saved = fmt_tokens(r.estimated_tokens_saved);
-    out.push_str(&format!(
-        "Without cartog  {without:>7} tokens   (~{TOKENS_PER_QUERY_GREP} / query)\n"
-    ));
-    out.push_str(&format!(
-        "With cartog     {with:>7} tokens   (~{TOKENS_PER_QUERY_CARTOG} / query)\n"
-    ));
-    out.push_str("──────────────────────────────────────────────\n");
-    out.push_str(&format!(
-        "Saved           {saved:>7} tokens   (~{} / query)\n\n",
-        r.baseline_delta
-    ));
-
-    // Per-tool counts with annotation so the reader knows the numbers are
-    // call counts, not per-tool token savings (the multiplier is flat).
-    out.push_str("By tool (call counts):\n");
-    for (tool, count) in &r.by_tool {
-        out.push_str(&format!("  {count:>4}  {tool}\n"));
-    }
-
-    if r.by_source.len() > 1 {
-        out.push_str("\nBy source:\n");
-        for (source, count) in &r.by_source {
-            out.push_str(&format!("  {count:>4}  {source}\n"));
-        }
-    }
-
-    out.push_str(&format!(
-        "\nBaseline: ~{TOKENS_PER_QUERY_GREP} tokens for an equivalent grep+read sweep \
-         vs cartog's ~{TOKENS_PER_QUERY_CARTOG}.\n\
-         Measured across 13 benchmark scenarios (see crates/cartog/benches/queries.rs).\n"
-    ));
-
-    out
-}
+mod savings;
+use savings::{render_savings, savings_scope_label};
 
 /// Index statistics summary.
 pub fn cmd_stats(
@@ -1323,286 +1214,8 @@ pub fn cmd_rag_search(
     })
 }
 
-/// Display the current configuration with default-value indicators.
-pub fn cmd_config(
-    config: &CartogConfig,
-    config_path: Option<&Path>,
-    config_rejected: bool,
-    db_path: &Path,
-    json: bool,
-) -> Result<()> {
-    // When the config file was found but rejected at parse time, `config`
-    // is the empty default — displaying it as the active config would
-    // silently lie. Show an explicit error and bail with the path so the
-    // user knows which file to fix.
-    if config_rejected {
-        // Invariant: ConfigLoad::Rejected always carries a path, and main
-        // propagates it as `config_path`. Bail explicitly so this surfaces
-        // even if the invariant ever changes.
-        let p = config_path
-            .ok_or_else(|| anyhow::anyhow!("config rejected but path missing — invariant break"))?;
-        anyhow::bail!(
-            "configuration file {} was rejected (see earlier stderr for the \
-             underlying reason). `cartog config` cannot display a meaningful \
-             view until the file is fixed.",
-            p.display()
-        );
-    }
-    use crate::config::{
-        DEFAULT_EMBEDDING_PROVIDER, DEFAULT_OLLAMA_BASE_URL, DEFAULT_OLLAMA_MODEL,
-        DEFAULT_RERANKER_PROVIDER,
-    };
-
-    let embed = config.embedding.as_ref();
-    let ollama = embed.and_then(|e| e.ollama.as_ref());
-    let local = embed.and_then(|e| e.local.as_ref());
-    let reranker = config.reranker.as_ref();
-    let rag = config.rag.as_ref();
-    let tuning_defaults = cartog_rag::search::SearchTuning::default();
-    // `to_search_tuning()` applies the clamps (retrieval_multiplier.max(1),
-    // rerank_min.min(rerank_max)) so what we show matches what the search
-    // pipeline will actually use.
-    let effective_tuning = rag.map(|r| r.to_search_tuning()).unwrap_or(tuning_defaults);
-
-    let rag_value = |set: Option<u32>, effective: u32, default: u32| ValueDisplay {
-        value: effective.to_string(),
-        is_default: set.is_none(),
-        default: default.to_string(),
-    };
-
-    let display = ConfigDisplay {
-        config_file: config_path.map(|p| p.to_string_lossy().into_owned()),
-        db_path: db_path.to_string_lossy().into_owned(),
-        embedding: EmbeddingDisplay {
-            provider: ValueDisplay {
-                value: embed.map_or(DEFAULT_EMBEDDING_PROVIDER.into(), |e| {
-                    e.provider().to_string()
-                }),
-                is_default: embed.map_or(true, |e| e.provider.is_none()),
-                default: DEFAULT_EMBEDDING_PROVIDER.into(),
-            },
-            model: embed.and_then(|e| e.model.clone()),
-            dimension: embed.and_then(|e| e.dimension),
-            local: LocalEmbeddingDisplay {
-                query_prefix: local.and_then(|l| l.query_prefix.clone()),
-                document_prefix: local.and_then(|l| l.document_prefix.clone()),
-            },
-            ollama: OllamaDisplay {
-                base_url: ValueDisplay {
-                    value: ollama
-                        .map_or(DEFAULT_OLLAMA_BASE_URL.into(), |o| o.base_url().to_string()),
-                    is_default: ollama.map_or(true, |o| o.base_url.is_none()),
-                    default: DEFAULT_OLLAMA_BASE_URL.into(),
-                },
-                model: ValueDisplay {
-                    value: ollama.map_or(DEFAULT_OLLAMA_MODEL.into(), |o| o.model().to_string()),
-                    is_default: ollama.map_or(true, |o| o.model.is_none()),
-                    default: DEFAULT_OLLAMA_MODEL.into(),
-                },
-            },
-        },
-        reranker: RerankerDisplay {
-            provider: ValueDisplay {
-                value: reranker.map_or(DEFAULT_RERANKER_PROVIDER.into(), |r| {
-                    r.provider().to_string()
-                }),
-                is_default: reranker.map_or(true, |r| r.provider.is_none()),
-                default: DEFAULT_RERANKER_PROVIDER.into(),
-            },
-        },
-        rag: RagDisplay {
-            retrieval_multiplier: rag_value(
-                rag.and_then(|r| r.retrieval_multiplier),
-                effective_tuning.retrieval_multiplier,
-                tuning_defaults.retrieval_multiplier,
-            ),
-            retrieval_floor: rag_value(
-                rag.and_then(|r| r.retrieval_floor),
-                effective_tuning.retrieval_floor,
-                tuning_defaults.retrieval_floor,
-            ),
-            rerank_max: rag_value(
-                rag.and_then(|r| r.rerank_max),
-                effective_tuning.rerank_max,
-                tuning_defaults.rerank_max,
-            ),
-            rerank_min: rag_value(
-                rag.and_then(|r| r.rerank_min),
-                effective_tuning.rerank_min,
-                tuning_defaults.rerank_min,
-            ),
-        },
-    };
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&display)?);
-    } else {
-        print!("{}", format_config_human(&display));
-    }
-    Ok(())
-}
-
-fn format_value(v: &ValueDisplay) -> String {
-    if v.is_default {
-        format!("{} (default)", v.value)
-    } else {
-        format!("{} (default: {})", v.value, v.default)
-    }
-}
-
-fn format_optional(v: &Option<String>) -> &str {
-    match v {
-        Some(s) => s.as_str(),
-        None => "-",
-    }
-}
-
-fn format_config_human(d: &ConfigDisplay) -> String {
-    use std::fmt::Write;
-    let mut out = String::new();
-
-    writeln!(
-        out,
-        "Config file: {}",
-        d.config_file.as_deref().unwrap_or("none")
-    )
-    .unwrap();
-    writeln!(out, "Database:    {}", d.db_path).unwrap();
-
-    writeln!(out, "\n[embedding]").unwrap();
-    writeln!(
-        out,
-        "  provider:          {}",
-        format_value(&d.embedding.provider)
-    )
-    .unwrap();
-    writeln!(
-        out,
-        "  model:             {}",
-        format_optional(&d.embedding.model)
-    )
-    .unwrap();
-    writeln!(
-        out,
-        "  dimension:         {}",
-        d.embedding.dimension.map_or("-".into(), |v| v.to_string())
-    )
-    .unwrap();
-
-    writeln!(out, "\n[embedding.local]").unwrap();
-    writeln!(
-        out,
-        "  query_prefix:      {}",
-        format_optional(&d.embedding.local.query_prefix)
-    )
-    .unwrap();
-    writeln!(
-        out,
-        "  document_prefix:   {}",
-        format_optional(&d.embedding.local.document_prefix)
-    )
-    .unwrap();
-
-    writeln!(out, "\n[embedding.ollama]").unwrap();
-    writeln!(
-        out,
-        "  base_url:          {}",
-        format_value(&d.embedding.ollama.base_url)
-    )
-    .unwrap();
-    writeln!(
-        out,
-        "  model:             {}",
-        format_value(&d.embedding.ollama.model)
-    )
-    .unwrap();
-
-    writeln!(out, "\n[reranker]").unwrap();
-    writeln!(
-        out,
-        "  provider:          {}",
-        format_value(&d.reranker.provider)
-    )
-    .unwrap();
-
-    writeln!(out, "\n[rag]").unwrap();
-    writeln!(
-        out,
-        "  retrieval_multiplier: {}",
-        format_value(&d.rag.retrieval_multiplier)
-    )
-    .unwrap();
-    writeln!(
-        out,
-        "  retrieval_floor:      {}",
-        format_value(&d.rag.retrieval_floor)
-    )
-    .unwrap();
-    writeln!(
-        out,
-        "  rerank_max:           {}",
-        format_value(&d.rag.rerank_max)
-    )
-    .unwrap();
-    writeln!(
-        out,
-        "  rerank_min:           {}",
-        format_value(&d.rag.rerank_min)
-    )
-    .unwrap();
-
-    out
-}
-
-#[derive(Serialize)]
-struct ConfigDisplay {
-    config_file: Option<String>,
-    db_path: String,
-    embedding: EmbeddingDisplay,
-    reranker: RerankerDisplay,
-    rag: RagDisplay,
-}
-
-#[derive(Serialize)]
-struct RagDisplay {
-    retrieval_multiplier: ValueDisplay,
-    retrieval_floor: ValueDisplay,
-    rerank_max: ValueDisplay,
-    rerank_min: ValueDisplay,
-}
-
-#[derive(Serialize)]
-struct EmbeddingDisplay {
-    provider: ValueDisplay,
-    model: Option<String>,
-    dimension: Option<usize>,
-    local: LocalEmbeddingDisplay,
-    ollama: OllamaDisplay,
-}
-
-#[derive(Serialize)]
-struct LocalEmbeddingDisplay {
-    query_prefix: Option<String>,
-    document_prefix: Option<String>,
-}
-
-#[derive(Serialize)]
-struct OllamaDisplay {
-    base_url: ValueDisplay,
-    model: ValueDisplay,
-}
-
-#[derive(Serialize)]
-struct RerankerDisplay {
-    provider: ValueDisplay,
-}
-
-#[derive(Serialize)]
-struct ValueDisplay {
-    value: String,
-    is_default: bool,
-    default: String,
-}
+mod config_display;
+pub use config_display::cmd_config;
 
 mod doctor;
 pub use doctor::cmd_doctor;
@@ -1645,148 +1258,6 @@ pub use self_cmd::{
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    // ── savings formatter helpers ─────────────────────────────────────
-
-    #[test]
-    fn fmt_tokens_under_thousand_renders_raw_integer() {
-        assert_eq!(fmt_tokens(0), "0");
-        assert_eq!(fmt_tokens(1), "1");
-        assert_eq!(fmt_tokens(999), "999");
-    }
-
-    #[test]
-    fn fmt_tokens_thousands_renders_one_decimal_k() {
-        assert_eq!(fmt_tokens(1_000), "~1.0k");
-        assert_eq!(fmt_tokens(1_420), "~1.4k");
-        assert_eq!(fmt_tokens(8_500), "~8.5k");
-        assert_eq!(fmt_tokens(999_999), "~1000.0k");
-    }
-
-    #[test]
-    fn fmt_tokens_millions_renders_one_decimal_m() {
-        assert_eq!(fmt_tokens(1_000_000), "~1.0M");
-        assert_eq!(fmt_tokens(2_500_000), "~2.5M");
-    }
-
-    #[test]
-    fn fmt_tokens_handles_u64_max_without_panicking() {
-        // Cosmetic — f64 loses precision near u64::MAX, but the call must
-        // not overflow or panic. Anchor a substring to catch obvious breakage.
-        let s = fmt_tokens(u64::MAX);
-        assert!(s.starts_with("~"));
-        assert!(s.ends_with("M"));
-    }
-
-    #[test]
-    fn savings_scope_label_handles_cartog_dir_layout() {
-        // The canonical layout: `<project>/.cartog/db.sqlite` — walk past
-        // `.cartog` to get the project dir.
-        assert_eq!(
-            savings_scope_label(Path::new("/home/user/myproject/.cartog/db.sqlite")),
-            "myproject"
-        );
-    }
-
-    #[test]
-    fn savings_scope_label_handles_legacy_layout() {
-        // The legacy layout: `<project>/.cartog.db` directly under the
-        // project root — parent is the project dir.
-        assert_eq!(
-            savings_scope_label(Path::new("/home/user/myproject/.cartog.db")),
-            "myproject"
-        );
-    }
-
-    #[test]
-    fn savings_scope_label_empty_on_unrecognizable_path() {
-        // Bare filename with no parent. render_savings drops the segment.
-        assert_eq!(savings_scope_label(Path::new("db.sqlite")), "");
-        // Root path — no parent above `.cartog`.
-        assert_eq!(savings_scope_label(Path::new("/.cartog/db.sqlite")), "");
-    }
-
-    #[test]
-    fn render_savings_zero_queries_emits_helpful_hint() {
-        let report = cartog_db::SavingsReport {
-            by_tool: Vec::new(),
-            by_source: Vec::new(),
-            total_queries: 0,
-            tokens_used_cartog: 0,
-            tokens_used_grep: 0,
-            estimated_tokens_saved: 0,
-            percent_saved: 0,
-            baseline_delta: cartog_db::TOKENS_SAVED_PER_QUERY,
-        };
-        let out = render_savings("myproject", &report);
-        assert!(out.contains("No queries logged yet"));
-        assert!(out.contains("cartog search"));
-    }
-
-    #[test]
-    fn render_savings_includes_header_bar_and_breakdown() {
-        let report = cartog_db::SavingsReport {
-            by_tool: vec![("search".to_string(), 3), ("refs".to_string(), 1)],
-            by_source: vec![("cli".to_string(), 4)],
-            total_queries: 4,
-            tokens_used_cartog: 1_120,
-            tokens_used_grep: 6_800,
-            estimated_tokens_saved: 5_680,
-            percent_saved: 83,
-            baseline_delta: cartog_db::TOKENS_SAVED_PER_QUERY,
-        };
-        let out = render_savings("myproject", &report);
-        // Header carries scope + query count.
-        assert!(out.contains("cartog · myproject · 4 queries"));
-        // ~83% → round-to-nearest gives 8 cells filled, 2 empty.
-        assert!(out.contains("████████░░"));
-        assert!(out.contains("~83% tokens saved"));
-        // With / without / saved rows present and formatted.
-        assert!(out.contains("Without cartog"));
-        assert!(out.contains("With cartog"));
-        assert!(out.contains("Saved"));
-        // Per-tool block labelled as call counts (no per-tool token figures).
-        assert!(out.contains("By tool (call counts):"));
-        assert!(out.contains("3  search"));
-    }
-
-    #[test]
-    fn render_savings_bar_at_99_percent_shows_full_ten_cells() {
-        // Regression for F4: previously `99 / 10 = 9` made 90% and 99%
-        // indistinguishable. Round-to-nearest fixes this.
-        let report = cartog_db::SavingsReport {
-            by_tool: vec![("search".into(), 1)],
-            by_source: vec![("cli".into(), 1)],
-            total_queries: 1,
-            tokens_used_cartog: 17,
-            tokens_used_grep: 1_700,
-            estimated_tokens_saved: 1_683,
-            percent_saved: 99,
-            baseline_delta: 1_420,
-        };
-        let out = render_savings("p", &report);
-        assert!(
-            out.contains("██████████"),
-            "99% must render 10 filled cells, got: {out}"
-        );
-    }
-
-    #[test]
-    fn render_savings_empty_scope_drops_label_segment() {
-        let report = cartog_db::SavingsReport {
-            by_tool: vec![("search".into(), 1)],
-            by_source: vec![("cli".into(), 1)],
-            total_queries: 1,
-            tokens_used_cartog: 280,
-            tokens_used_grep: 1_700,
-            estimated_tokens_saved: 1_420,
-            percent_saved: 83,
-            baseline_delta: 1_420,
-        };
-        let out = render_savings("", &report);
-        // Falls back to `cartog · N queries` without the scope segment.
-        assert!(out.starts_with("cartog · 1 queries"), "got: {out}");
-    }
 
     #[test]
     fn capitalized_index_phase_labels() {
@@ -1921,140 +1392,6 @@ mod tests {
         let text = "abcd"; // 4 bytes = 1 token
         let result = truncate_to_budget(text, 1);
         assert_eq!(result, "abcd");
-    }
-
-    // ── Config display tests ──
-
-    fn default_config_display() -> ConfigDisplay {
-        use crate::config::{
-            DEFAULT_EMBEDDING_PROVIDER, DEFAULT_OLLAMA_BASE_URL, DEFAULT_OLLAMA_MODEL,
-            DEFAULT_RERANKER_PROVIDER,
-        };
-        ConfigDisplay {
-            config_file: None,
-            db_path: "/tmp/test.db".into(),
-            embedding: EmbeddingDisplay {
-                provider: ValueDisplay {
-                    value: DEFAULT_EMBEDDING_PROVIDER.into(),
-                    is_default: true,
-                    default: DEFAULT_EMBEDDING_PROVIDER.into(),
-                },
-                model: None,
-                dimension: None,
-                local: LocalEmbeddingDisplay {
-                    query_prefix: None,
-                    document_prefix: None,
-                },
-                ollama: OllamaDisplay {
-                    base_url: ValueDisplay {
-                        value: DEFAULT_OLLAMA_BASE_URL.into(),
-                        is_default: true,
-                        default: DEFAULT_OLLAMA_BASE_URL.into(),
-                    },
-                    model: ValueDisplay {
-                        value: DEFAULT_OLLAMA_MODEL.into(),
-                        is_default: true,
-                        default: DEFAULT_OLLAMA_MODEL.into(),
-                    },
-                },
-            },
-            reranker: RerankerDisplay {
-                provider: ValueDisplay {
-                    value: DEFAULT_RERANKER_PROVIDER.into(),
-                    is_default: true,
-                    default: DEFAULT_RERANKER_PROVIDER.into(),
-                },
-            },
-            rag: {
-                let t = cartog_rag::search::SearchTuning::default();
-                let v = |n: u32| ValueDisplay {
-                    value: n.to_string(),
-                    is_default: true,
-                    default: n.to_string(),
-                };
-                RagDisplay {
-                    retrieval_multiplier: v(t.retrieval_multiplier),
-                    retrieval_floor: v(t.retrieval_floor),
-                    rerank_max: v(t.rerank_max),
-                    rerank_min: v(t.rerank_min),
-                }
-            },
-        }
-    }
-
-    #[test]
-    fn test_format_config_human_all_defaults() {
-        let d = default_config_display();
-        let out = format_config_human(&d);
-        assert!(out.contains("Config file: none"));
-        assert!(out.contains("Database:    /tmp/test.db"));
-        assert!(out.contains("local (default)"));
-        assert!(out.contains("model:             -"));
-        assert!(out.contains("dimension:         -"));
-        assert!(out.contains("query_prefix:      -"));
-        assert!(out.contains("document_prefix:   -"));
-    }
-
-    #[test]
-    fn test_format_config_human_custom_values() {
-        let mut d = default_config_display();
-        d.config_file = Some("/project/.cartog.toml".into());
-        d.embedding.provider = ValueDisplay {
-            value: "ollama".into(),
-            is_default: false,
-            default: "local".into(),
-        };
-        d.embedding.model = Some("nomic-embed-text".into());
-        d.embedding.dimension = Some(768);
-
-        let out = format_config_human(&d);
-        assert!(out.contains("Config file: /project/.cartog.toml"));
-        assert!(out.contains("ollama (default: local)"));
-        assert!(out.contains("model:             nomic-embed-text"));
-        assert!(out.contains("dimension:         768"));
-    }
-
-    #[test]
-    fn test_format_value_default() {
-        let v = ValueDisplay {
-            value: "local".into(),
-            is_default: true,
-            default: "local".into(),
-        };
-        assert_eq!(format_value(&v), "local (default)");
-    }
-
-    #[test]
-    fn test_format_value_overridden() {
-        let v = ValueDisplay {
-            value: "ollama".into(),
-            is_default: false,
-            default: "local".into(),
-        };
-        assert_eq!(format_value(&v), "ollama (default: local)");
-    }
-
-    #[test]
-    fn test_format_optional_some() {
-        let v = Some("value".to_string());
-        assert_eq!(format_optional(&v), "value");
-    }
-
-    #[test]
-    fn test_format_optional_none() {
-        let v: Option<String> = None;
-        assert_eq!(format_optional(&v), "-");
-    }
-
-    #[test]
-    fn test_config_display_json_serialization() {
-        let d = default_config_display();
-        let json = serde_json::to_value(&d).unwrap();
-        assert_eq!(json["db_path"], "/tmp/test.db");
-        assert_eq!(json["embedding"]["provider"]["value"], "local");
-        assert_eq!(json["embedding"]["provider"]["is_default"], true);
-        assert!(json["config_file"].is_null());
-        assert!(json["embedding"]["model"].is_null());
     }
 
     #[test]
